move more utilites into test_coprocessor
