+++ tests/coprocessor/test_select.rs
@@ -11,314 +11,34 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::cmp;
 use std::i64;
 use std::thread;
-use std::{cmp, mem};
 
-use protobuf::{Message, RepeatedField};
+use protobuf::Message;
 
-use kvproto::coprocessor::{KeyRange, Request, Response};
+use kvproto::coprocessor::Response;
 use kvproto::kvrpcpb::Context;
-use tipb::executor::{
-    Aggregation, ExecType, Executor, IndexScan, Limit, Selection, TableScan, TopN,
-};
-use tipb::expression::{ByItem, Expr, ExprType, ScalarFuncSig};
-use tipb::schema::ColumnInfo;
-use tipb::select::{Chunk, DAGRequest, EncodeType};
+use tipb::expression::{Expr, ExprType, ScalarFuncSig};
+use tipb::select::{Chunk, EncodeType};
 
 use test_coprocessor::*;
 use test_storage::*;
-use tikv::coprocessor::codec::{datum, table, Datum};
-use tikv::coprocessor::*;
+use tikv::coprocessor::codec::{datum, Datum};
+use tikv::server::readpool;
 use tikv::server::Config;
-use tikv::storage::engine::{self, RocksEngine, TEMP_DIR};
+use tikv::storage::engine::{self, TEMP_DIR};
 use tikv::storage::ALL_CFS;
 use tikv::util::codec::number::*;
-use tikv::util::worker::Worker;
 
 const FLAG_IGNORE_TRUNCATE: u64 = 1;
 const FLAG_TRUNCATE_AS_WARNING: u64 = 1 << 1;
-
-struct DAGChunkSpliter {
-    chunks: Vec<Chunk>,
-    datums: Vec<Datum>,
-    col_cnt: usize,
-}
-
-impl DAGChunkSpliter {
-    fn new(chunks: Vec<Chunk>, col_cnt: usize) -> DAGChunkSpliter {
-        DAGChunkSpliter {
-            chunks,
-            col_cnt,
-            datums: Vec::with_capacity(0),
-        }
-    }
-}
-
-impl Iterator for DAGChunkSpliter {
-    type Item = Vec<Datum>;
-
-    fn next(&mut self) -> Option<Vec<Datum>> {
-        loop {
-            if self.chunks.is_empty() && self.datums.is_empty() {
-                return None;
-            } else if self.datums.is_empty() {
-                let chunk = self.chunks.remove(0);
-                let mut data = chunk.get_rows_data();
-                self.datums = datum::decode(&mut data).unwrap();
-                continue;
-            }
-            assert_eq!(self.datums.len() >= self.col_cnt, true);
-            let mut cols = self.datums.split_off(self.col_cnt);
-            mem::swap(&mut self.datums, &mut cols);
-            return Some(cols);
-        }
-    }
-}
 
 fn check_chunk_datum_count(chunks: &[Chunk], datum_limit: usize) {
     let mut iter = chunks.iter();
     let res = iter.any(|x| datum::decode(&mut x.get_rows_data()).unwrap().len() != datum_limit);
     if res {
         assert!(iter.next().is_none());
-    }
-}
-
-// This function will create a Product table and initialize with the specified data.
-fn init_with_data(
-    tbl: &ProductTable,
-    vals: &[(i64, Option<&str>, i64)],
-) -> (Store<RocksEngine>, Worker<EndPointTask<RocksEngine>>) {
-    init_data_with_commit(tbl, vals, true)
-}
-
-fn offset_for_column(cols: &[ColumnInfo], col_id: i64) -> i64 {
-    for (offset, column) in cols.iter().enumerate() {
-        if column.get_column_id() == col_id {
-            return offset as i64;
-        }
-    }
-    0 as i64
-}
-
-struct DAGSelect {
-    execs: Vec<Executor>,
-    cols: Vec<ColumnInfo>,
-    order_by: Vec<ByItem>,
-    limit: Option<u64>,
-    aggregate: Vec<Expr>,
-    group_by: Vec<Expr>,
-    key_range: KeyRange,
-    output_offsets: Option<Vec<u32>>,
-}
-
-impl DAGSelect {
-    fn from(table: &Table) -> DAGSelect {
-        let mut exec = Executor::new();
-        exec.set_tp(ExecType::TypeTableScan);
-        let mut tbl_scan = TableScan::new();
-        let mut table_info = table.get_table_info();
-        tbl_scan.set_table_id(table_info.get_table_id());
-        let columns_info = table_info.take_columns();
-        tbl_scan.set_columns(columns_info);
-        exec.set_tbl_scan(tbl_scan);
-
-        let mut range = KeyRange::new();
-        range.set_start(table::encode_row_key(table.id, i64::MIN));
-        range.set_end(table::encode_row_key(table.id, i64::MAX));
-
-        DAGSelect {
-            execs: vec![exec],
-            cols: table.get_table_columns(),
-            order_by: vec![],
-            limit: None,
-            aggregate: vec![],
-            group_by: vec![],
-            key_range: range,
-            output_offsets: None,
-        }
-    }
-
-    fn from_index(table: &Table, index: Column) -> DAGSelect {
-        let idx = index.index;
-        let mut exec = Executor::new();
-        exec.set_tp(ExecType::TypeIndexScan);
-        let mut scan = IndexScan::new();
-        let mut index_info = table.get_index_info(idx, true);
-        scan.set_table_id(index_info.get_table_id());
-        scan.set_index_id(idx);
-
-        let columns_info = index_info.take_columns();
-        scan.set_columns(columns_info.clone());
-        exec.set_idx_scan(scan);
-
-        let range = table.get_index_range(idx);
-        DAGSelect {
-            execs: vec![exec],
-            cols: columns_info.to_vec(),
-            order_by: vec![],
-            limit: None,
-            aggregate: vec![],
-            group_by: vec![],
-            key_range: range,
-            output_offsets: None,
-        }
-    }
-
-    fn limit(mut self, n: u64) -> DAGSelect {
-        self.limit = Some(n);
-        self
-    }
-
-    fn order_by(mut self, col: Column, desc: bool) -> DAGSelect {
-        let col_offset = offset_for_column(&self.cols, col.id);
-        let mut item = ByItem::new();
-        let mut expr = Expr::new();
-        expr.set_tp(ExprType::ColumnRef);
-        expr.mut_val().encode_i64(col_offset).unwrap();
-        item.set_expr(expr);
-        item.set_desc(desc);
-        self.order_by.push(item);
-        self
-    }
-
-    fn count(mut self) -> DAGSelect {
-        let mut expr = Expr::new();
-        expr.set_tp(ExprType::Count);
-        self.aggregate.push(expr);
-        self
-    }
-
-    fn aggr_col(mut self, col: Column, aggr_t: ExprType) -> DAGSelect {
-        let col_offset = offset_for_column(&self.cols, col.id);
-        let mut col_expr = Expr::new();
-        col_expr.set_tp(ExprType::ColumnRef);
-        col_expr.mut_val().encode_i64(col_offset).unwrap();
-        let mut expr = Expr::new();
-        expr.set_tp(aggr_t);
-        expr.mut_children().push(col_expr);
-        self.aggregate.push(expr);
-        self
-    }
-
-    fn first(self, col: Column) -> DAGSelect {
-        self.aggr_col(col, ExprType::First)
-    }
-
-    fn sum(self, col: Column) -> DAGSelect {
-        self.aggr_col(col, ExprType::Sum)
-    }
-
-    fn avg(self, col: Column) -> DAGSelect {
-        self.aggr_col(col, ExprType::Avg)
-    }
-
-    fn max(self, col: Column) -> DAGSelect {
-        self.aggr_col(col, ExprType::Max)
-    }
-
-    fn min(self, col: Column) -> DAGSelect {
-        self.aggr_col(col, ExprType::Min)
-    }
-
-    fn bit_and(self, col: Column) -> DAGSelect {
-        self.aggr_col(col, ExprType::Agg_BitAnd)
-    }
-
-    fn bit_or(self, col: Column) -> DAGSelect {
-        self.aggr_col(col, ExprType::Agg_BitOr)
-    }
-
-    fn bit_xor(self, col: Column) -> DAGSelect {
-        self.aggr_col(col, ExprType::Agg_BitXor)
-    }
-
-    fn group_by(mut self, cols: &[Column]) -> DAGSelect {
-        for col in cols {
-            let offset = offset_for_column(&self.cols, col.id);
-            let mut expr = Expr::new();
-            expr.set_tp(ExprType::ColumnRef);
-            expr.mut_val().encode_i64(offset).unwrap();
-            self.group_by.push(expr);
-        }
-        self
-    }
-
-    fn output_offsets(mut self, output_offsets: Option<Vec<u32>>) -> DAGSelect {
-        self.output_offsets = output_offsets;
-        self
-    }
-
-    fn where_expr(mut self, expr: Expr) -> DAGSelect {
-        let mut exec = Executor::new();
-        exec.set_tp(ExecType::TypeSelection);
-        let mut selection = Selection::new();
-        selection.mut_conditions().push(expr);
-        exec.set_selection(selection);
-        self.execs.push(exec);
-        self
-    }
-
-    fn build(self) -> Request {
-        self.build_with(Context::new(), &[0])
-    }
-
-    fn build_with(mut self, ctx: Context, flags: &[u64]) -> Request {
-        if !self.aggregate.is_empty() || !self.group_by.is_empty() {
-            let mut exec = Executor::new();
-            exec.set_tp(ExecType::TypeAggregation);
-            let mut aggr = Aggregation::new();
-            if !self.aggregate.is_empty() {
-                aggr.set_agg_func(RepeatedField::from_vec(self.aggregate));
-            }
-
-            if !self.group_by.is_empty() {
-                aggr.set_group_by(RepeatedField::from_vec(self.group_by));
-            }
-            exec.set_aggregation(aggr);
-            self.execs.push(exec);
-        }
-
-        if !self.order_by.is_empty() {
-            let mut exec = Executor::new();
-            exec.set_tp(ExecType::TypeTopN);
-            let mut topn = TopN::new();
-            topn.set_order_by(RepeatedField::from_vec(self.order_by));
-            if let Some(limit) = self.limit.take() {
-                topn.set_limit(limit);
-            }
-            exec.set_topN(topn);
-            self.execs.push(exec);
-        }
-
-        if let Some(l) = self.limit.take() {
-            let mut exec = Executor::new();
-            exec.set_tp(ExecType::TypeLimit);
-            let mut limit = Limit::new();
-            limit.set_limit(l);
-            exec.set_limit(limit);
-            self.execs.push(exec);
-        }
-
-        let mut dag = DAGRequest::new();
-        dag.set_executors(RepeatedField::from_vec(self.execs));
-        dag.set_start_ts(next_id() as u64);
-        dag.set_flags(flags.iter().fold(0, |acc, f| acc | *f));
-        dag.set_collect_range_counts(true);
-
-        let output_offsets = if self.output_offsets.is_some() {
-            self.output_offsets.take().unwrap()
-        } else {
-            (0..self.cols.len() as u32).collect()
-        };
-        dag.set_output_offsets(output_offsets);
-
-        let mut req = Request::new();
-        req.set_tp(REQ_TYPE_DAG);
-        req.set_data(dag.write_to_bytes().unwrap());
-        req.set_ranges(RepeatedField::from_vec(vec![self.key_range]));
-        req.set_context(ctx);
-        req
     }
 }
 
@@ -335,7 +55,7 @@ fn test_select() {
     let (_, mut end_point) = init_with_data(&product, &data);
     // for dag selection
     let req = DAGSelect::from(&product.table).build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 3);
     for (row, (id, name, cnt)) in spliter.zip(data) {
         let name_datum = name.map(|s| s.as_bytes()).into();
@@ -363,12 +83,20 @@ fn test_batch_row_limit() {
         let engine = engine::new_local_engine(TEMP_DIR, ALL_CFS).unwrap();
         let mut cfg = Config::default();
         cfg.end_point_batch_row_limit = batch_row_limit;
-        init_data_with_details(Context::new(), engine, &product, &data, true, cfg)
+        init_data_with_details(
+            Context::new(),
+            engine,
+            &product,
+            &data,
+            true,
+            &cfg,
+            &readpool::Config::default_for_test(),
+        )
     };
 
     // for dag selection
     let req = DAGSelect::from(&product.table).build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     check_chunk_datum_count(resp.get_chunks(), chunk_datum_limit);
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 3);
     for (row, (id, name, cnt)) in spliter.zip(data) {
@@ -398,7 +126,15 @@ fn test_stream_batch_row_limit() {
         let engine = engine::new_local_engine(TEMP_DIR, ALL_CFS).unwrap();
         let mut cfg = Config::default();
         cfg.end_point_stream_batch_row_limit = stream_row_limit;
-        init_data_with_details(Context::new(), engine, &product, &data, true, cfg)
+        init_data_with_details(
+            Context::new(),
+            engine,
+            &product,
+            &data,
+            true,
+            &cfg,
+            &readpool::Config::default_for_test(),
+        )
     };
 
     let req = DAGSelect::from(&product.table).build();
@@ -413,7 +149,7 @@ fn test_stream_batch_row_limit() {
         assert_eq!(end[end.len() - 1], end_last_byte);
     };
 
-    let resps = handle_streaming_select(&end_point, req, check_range);
+    let resps = handle_streaming_select(&end_point.scheduler(), req, check_range);
     assert_eq!(resps.len(), 3);
     let expected_output_counts = vec![vec![2 as i64], vec![2 as i64], vec![1 as i64]];
     for (i, resp) in resps.into_iter().enumerate() {
@@ -462,7 +198,7 @@ fn test_select_after_lease() {
     // Sleep until the leader lease is expired.
     thread::sleep(cluster.cfg.raft_store.raft_store_max_leader_lease.0);
     let req = DAGSelect::from(&product.table).build_with(ctx.clone(), &[0]);
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 3);
     for (row, (id, name, cnt)) in spliter.zip(data) {
         let name_datum = name.map(|s| s.as_bytes()).into();
@@ -489,7 +225,15 @@ fn test_scan_detail() {
         let engine = engine::new_local_engine(TEMP_DIR, ALL_CFS).unwrap();
         let mut cfg = Config::default();
         cfg.end_point_batch_row_limit = 50;
-        init_data_with_details(Context::new(), engine, &product, &data, true, cfg)
+        init_data_with_details(
+            Context::new(),
+            engine,
+            &product,
+            &data,
+            true,
+            &cfg,
+            &readpool::Config::default_for_test(),
+        )
     };
 
     let reqs = vec![
@@ -501,7 +245,7 @@ fn test_scan_detail() {
         req.mut_context().set_scan_detail(true);
         req.mut_context().set_handle_time(true);
 
-        let resp = handle_request(&end_point, req);
+        let resp = handle_request(&end_point.scheduler(), req);
         assert!(resp.get_exec_details().has_handle_time());
 
         let scan_detail = resp.get_exec_details().get_scan_detail();
@@ -529,7 +273,7 @@ fn test_group_by() {
     let req = DAGSelect::from(&product.table)
         .group_by(&[product.name])
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     // should only have name:0, name:2 and name:1
     let mut row_count = 0;
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 1);
@@ -569,7 +313,7 @@ fn test_aggr_count() {
         .count()
         .group_by(&[product.name])
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let exp_len = exp.len();
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 2);
@@ -595,7 +339,7 @@ fn test_aggr_count() {
         .count()
         .group_by(&[product.name, product.count])
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let exp_len = exp.len();
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 3);
@@ -642,7 +386,7 @@ fn test_aggr_first() {
         .first(product.id)
         .group_by(&[product.name])
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let exp_len = exp.len();
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 2);
@@ -669,7 +413,7 @@ fn test_aggr_first() {
         .first(product.name)
         .group_by(&[product.count])
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let exp_len = exp.len();
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 2);
@@ -720,7 +464,7 @@ fn test_aggr_avg() {
         .avg(product.count)
         .group_by(&[product.name])
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let exp_len = exp.len();
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 3);
@@ -761,7 +505,7 @@ fn test_aggr_sum() {
         .sum(product.count)
         .group_by(&[product.name])
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let exp_len = exp.len();
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 2);
@@ -827,7 +571,7 @@ fn test_aggr_extre() {
         .min(product.count)
         .group_by(&[product.name])
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let exp_len = exp.len();
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 3);
@@ -903,7 +647,7 @@ fn test_aggr_bit_ops() {
         .bit_xor(product.count)
         .group_by(&[product.name])
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let exp_len = exp.len();
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 4);
@@ -947,7 +691,7 @@ fn test_order_by_column() {
         .order_by(product.name, false)
         .limit(5)
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 3);
     for (row, (id, name, cnt)) in spliter.zip(exp) {
@@ -983,7 +727,7 @@ fn test_order_by_pk_with_select_from_index() {
         .order_by(product.id, true)
         .limit(5)
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 3);
     for (row, (id, name, cnt)) in spliter.zip(expect) {
@@ -1014,7 +758,7 @@ fn test_limit() {
     let expect: Vec<_> = data.drain(..5).collect();
     // for dag
     let req = DAGSelect::from(&product.table).limit(5).build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 3);
     for (row, (id, name, cnt)) in spliter.zip(expect) {
@@ -1049,7 +793,7 @@ fn test_reverse() {
         .limit(5)
         .order_by(product.id, true)
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 3);
     for (row, (id, name, cnt)) in spliter.zip(expect) {
@@ -1079,7 +823,7 @@ fn test_index() {
     let (_, mut end_point) = init_with_data(&product, &data);
     // for dag
     let req = DAGSelect::from_index(&product.table, product.id).build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 1);
     for (row, (id, _, _)) in spliter.zip(data) {
@@ -1114,7 +858,7 @@ fn test_index_reverse_limit() {
         .order_by(product.id, true)
         .build();
 
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 1);
     for (row, (id, _, _)) in spliter.zip(expect) {
@@ -1145,7 +889,7 @@ fn test_limit_oom() {
     let req = DAGSelect::from_index(&product.table, product.id)
         .limit(100000000)
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 1);
     for (row, (id, _, _)) in spliter.zip(data) {
@@ -1182,7 +926,7 @@ fn test_del_select() {
 
     // for dag
     let req = DAGSelect::from_index(&product.table, product.id).build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 1);
     let mut row_count = 0;
     for _ in spliter {
@@ -1208,7 +952,7 @@ fn test_index_group_by() {
     let req = DAGSelect::from_index(&product.table, product.name)
         .group_by(&[product.name])
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     // should only have name:0, name:2 and name:1
     let mut row_count = 0;
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 1);
@@ -1240,7 +984,7 @@ fn test_index_aggr_count() {
     let req = DAGSelect::from_index(&product.table, product.name)
         .count()
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 1);
     let expected_encoded = datum::encode_value(&[Datum::U64(data.len() as u64)]).unwrap();
     let ret_data = spliter.next();
@@ -1260,7 +1004,7 @@ fn test_index_aggr_count() {
         .count()
         .group_by(&[product.name])
         .build();
-    resp = handle_select(&end_point, req);
+    resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let exp_len = exp.len();
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 2);
@@ -1284,7 +1028,7 @@ fn test_index_aggr_count() {
         .count()
         .group_by(&[product.name, product.count])
         .build();
-    resp = handle_select(&end_point, req);
+    resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let exp_len = exp.len();
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 3);
@@ -1326,7 +1070,7 @@ fn test_index_aggr_first() {
         .first(product.id)
         .group_by(&[product.name])
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let exp_len = exp.len();
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 2);
@@ -1377,7 +1121,7 @@ fn test_index_aggr_avg() {
         .avg(product.count)
         .group_by(&[product.name])
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let exp_len = exp.len();
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 3);
@@ -1417,7 +1161,7 @@ fn test_index_aggr_sum() {
         .sum(product.count)
         .group_by(&[product.name])
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let exp_len = exp.len();
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 2);
@@ -1482,7 +1226,7 @@ fn test_index_aggr_extre() {
         .min(product.count)
         .group_by(&[product.name])
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let exp_len = exp.len();
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 3);
@@ -1532,7 +1276,7 @@ fn test_where() {
     };
 
     let req = DAGSelect::from(&product.table).where_expr(cond).build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 3);
     let row = spliter.next().unwrap();
     let (id, name, cnt) = data[2];
@@ -1627,7 +1371,7 @@ fn test_handle_truncate() {
         let req = DAGSelect::from(&product.table)
             .where_expr(cond.clone())
             .build_with(Context::new(), &[FLAG_IGNORE_TRUNCATE]);
-        let resp = handle_select(&end_point, req);
+        let resp = handle_select(&end_point.scheduler(), req);
         assert!(!resp.has_error());
         assert!(resp.get_warnings().is_empty());
 
@@ -1635,7 +1379,7 @@ fn test_handle_truncate() {
         let req = DAGSelect::from(&product.table)
             .where_expr(cond.clone())
             .build_with(Context::new(), &[FLAG_TRUNCATE_AS_WARNING]);
-        let mut resp = handle_select(&end_point, req);
+        let mut resp = handle_select(&end_point.scheduler(), req);
         assert!(!resp.has_error());
         assert!(!resp.get_warnings().is_empty());
         // check data
@@ -1653,7 +1397,7 @@ fn test_handle_truncate() {
         let req = DAGSelect::from(&product.table)
             .where_expr(cond.clone())
             .build();
-        let mut resp = handle_select(&end_point, req);
+        let mut resp = handle_select(&end_point.scheduler(), req);
         assert!(resp.has_error());
         assert!(resp.get_warnings().is_empty());
     }
@@ -1685,7 +1429,7 @@ fn test_default_val() {
     let (_, mut end_point) = init_with_data(&product, &data);
     let expect: Vec<_> = data.drain(..5).collect();
     let req = DAGSelect::from(&tbl).limit(5).build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let mut row_count = 0;
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 4);
     for (row, (id, name, cnt)) in spliter.zip(expect) {
@@ -1716,7 +1460,7 @@ fn test_output_offsets() {
     let req = DAGSelect::from(&product.table)
         .output_offsets(Some(vec![1]))
         .build();
-    let mut resp = handle_select(&end_point, req);
+    let mut resp = handle_select(&end_point.scheduler(), req);
     let spliter = DAGChunkSpliter::new(resp.take_chunks().into_vec(), 1);
     for (row, (_, name, _)) in spliter.zip(data) {
         let name_datum = name.map(|s| s.as_bytes()).into();
@@ -1741,7 +1485,7 @@ fn test_key_is_locked_for_primary() {
     let (_, mut end_point) = init_data_with_commit(&product, &data, false);
 
     let req = DAGSelect::from(&product.table).build();
-    let resp = handle_request(&end_point, req);
+    let resp = handle_request(&end_point.scheduler(), req);
     assert!(resp.get_data().is_empty(), "{:?}", resp);
     assert!(resp.has_locked(), "{:?}", resp);
     end_point.stop().unwrap().join().unwrap();
@@ -1760,7 +1504,7 @@ fn test_key_is_locked_for_index() {
     let (_, mut end_point) = init_data_with_commit(&product, &data, false);
 
     let req = DAGSelect::from_index(&product.table, product.name).build();
-    let resp = handle_request(&end_point, req);
+    let resp = handle_request(&end_point.scheduler(), req);
     assert!(resp.get_data().is_empty(), "{:?}", resp);
     assert!(resp.has_locked(), "{:?}", resp);
     end_point.stop().unwrap().join().unwrap();
@@ -1779,7 +1523,7 @@ fn test_output_counts() {
     let (_, mut end_point) = init_with_data(&product, &data);
 
     let req = DAGSelect::from(&product.table).build();
-    let resp = handle_select(&end_point, req);
+    let resp = handle_select(&end_point.scheduler(), req);
     assert_eq!(resp.get_output_counts(), [data.len() as i64]);
 
     end_point.stop().unwrap().join().unwrap();
@@ -1799,7 +1543,7 @@ fn test_exec_details() {
 
     // get none
     let req = DAGSelect::from(&product.table).build();
-    let resp = handle_request(&end_point, req);
+    let resp = handle_request(&end_point.scheduler(), req);
     assert!(!resp.has_exec_details());
 
     let flags = &[0];
@@ -1808,7 +1552,7 @@ fn test_exec_details() {
     let mut ctx = Context::new();
     ctx.set_handle_time(true);
     let req = DAGSelect::from(&product.table).build_with(ctx, flags);
-    let resp = handle_request(&end_point, req);
+    let resp = handle_request(&end_point.scheduler(), req);
     assert!(resp.has_exec_details());
     let exec_details = resp.get_exec_details();
     assert!(exec_details.has_handle_time());
@@ -1818,7 +1562,7 @@ fn test_exec_details() {
     let mut ctx = Context::new();
     ctx.set_scan_detail(true);
     let req = DAGSelect::from(&product.table).build_with(ctx, flags);
-    let resp = handle_request(&end_point, req);
+    let resp = handle_request(&end_point.scheduler(), req);
     assert!(resp.has_exec_details());
     let exec_details = resp.get_exec_details();
     assert!(!exec_details.has_handle_time());
@@ -1829,7 +1573,7 @@ fn test_exec_details() {
     ctx.set_scan_detail(true);
     ctx.set_handle_time(true);
     let req = DAGSelect::from(&product.table).build_with(ctx, flags);
-    let resp = handle_request(&end_point, req);
+    let resp = handle_request(&end_point.scheduler(), req);
     assert!(resp.has_exec_details());
     let exec_details = resp.get_exec_details();
     assert!(exec_details.has_handle_time());
@@ -1854,7 +1598,7 @@ fn test_invalid_range() {
     select.key_range.set_start(b"xxx".to_vec());
     select.key_range.set_end(b"zzz".to_vec());
     let req = select.build();
-    let resp = handle_request(&end_point, req);
+    let resp = handle_request(&end_point.scheduler(), req);
     assert!(!resp.get_other_error().is_empty());
 
     end_point.stop().unwrap().join().unwrap();
