+++ tests/failpoints.rs
@@ -24,6 +24,7 @@ extern crate lazy_static;
 #[macro_use]
 extern crate log;
 
+extern crate test_coprocessor;
 extern crate test_raftstore;
 extern crate test_storage;
 extern crate test_util;
